Fix ansible user undefined
I had been using a variable set in the inventory to determine the current user
this did not work correctly when running this role in my devbox playbook.
The solution was to 'assume' that the playbook is running in the users home
directory and to not use a 'full' path to the git repo folder, but to use
a relative path from the current users home directory.